+++ code/parachain/frame/pablo/src/dual_asset_constant_product.rs
@@ -1,12 +1,10 @@
-use core::ops::Sub;
-
 use crate::{Config, Error, PoolConfiguration, PoolCount, Pools};
 use composable_maths::dex::{
 	constant_product::{
 		compute_deposit_lp_, compute_first_deposit_lp_, compute_in_given_out_new,
 		compute_out_given_in_new, compute_redeemed_for_lp,
 	},
-	PoolWeightMathExt,
+	per_thing_acceptable_computation_error, PoolWeightMathExt,
 };
 use composable_support::math::safe::SafeAdd;
 use composable_traits::{
@@ -20,7 +18,7 @@ use frame_support::{
 };
 use sp_runtime::{
 	traits::{Convert, One, Zero},
-	BoundedBTreeMap, PerThing, Permill,
+	BoundedBTreeMap, Permill,
 };
 use sp_std::{collections::btree_map::BTreeMap, vec::Vec};
 
@@ -100,7 +98,7 @@ impl<T: Config> DualAssetConstantProduct<T> {
 				};
 
 				let balance =
-					pool_assets.remove(&asset_amount.asset_id).ok_or(Error::<T>::PairMismatch)?;
+					pool_assets.remove(&asset_amount.asset_id).ok_or(Error::<T>::AssetNotFound)?;
 
 				Ok((asset_amount, balance))
 			})
@@ -156,6 +154,8 @@ impl<T: Config> DualAssetConstantProduct<T> {
 						Error::<T>::IncorrectAmountOfAssets
 					);
 
+					// pass 1 as weight since adding liquidity for all assets
+					// see docs on compute_deposit_lp_ for more information
 					let first_deposit = compute_deposit_lp_(
 						lp_total_issuance,
 						T::Convert::convert(first.amount),
@@ -221,7 +221,7 @@ impl<T: Config> DualAssetConstantProduct<T> {
 				};
 
 				let balance =
-					pool_assets.remove(&asset_amount.asset_id).ok_or(Error::<T>::PairMismatch)?;
+					pool_assets.remove(&asset_amount.asset_id).ok_or(Error::<T>::AssetNotFound)?;
 
 				Ok((asset_amount, balance))
 			})
@@ -366,46 +366,5 @@ impl<T: Config> DualAssetConstantProduct<T> {
 		let fee = pool.fee_config.calculate_fees(in_asset_id, T::Convert::convert(amm_pair.fee));
 
 		Ok((out_asset, a_sent, fee))
-	}
-}
-
-fn per_thing_acceptable_computation_error<T>(a: T, b: T) -> bool
-where
-	T: PerThing + Sub<Output = T>,
-{
-	let c = if a > b { a - b } else { b - a };
-
-	let epsilon = T::from_rational::<u128>(1, 100_000);
-
-	c <= epsilon
-}
-
-#[cfg(test)]
-mod per_thing_computation_error {
-	use sp_arithmetic::Rounding;
-	use sp_runtime::{PerThing, Permill};
-
-	use super::per_thing_acceptable_computation_error;
-
-	#[test]
-	fn difference_too_large() {
-		let a = Permill::from_rational::<u32>(1, 100_000);
-		let b = Permill::from_rational::<u32>(3, 100_000);
-
-		assert!(!per_thing_acceptable_computation_error(a, b))
-	}
-
-	#[test]
-	fn acceptable_difference() {
-		let a = Permill::from_rational_with_rounding::<u128>(
-			49_999_000,
-			100_000_000,
-			Rounding::NearestPrefUp,
-		)
-		.unwrap();
-		// let a = Permill::from_rational::<u128>(50001, 100_000);
-		let b = Permill::from_rational::<u32>(50, 100);
-
-		assert!(per_thing_acceptable_computation_error(a, b))
 	}
 }
